fix: moved classes
